requirements.txt
certifi==2020.6.20
cycler==0.10.0
gif==2.0.0
kiwisolver==1.2.0
matplotlib==3.3.0
numpy==1.19.0
opencv-python==4.3.0.36
pandas==1.0.5
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
scipy==1.5.1
seaborn==0.10.1
six==1.15.0
